+++ rdf-delta-server-http/src/main/java/org/seaborne/delta/server/http/S_Patch.java
@@ -86,7 +86,7 @@ public class S_Patch extends HttpOperationBase {
             if ( false )
                 RDFPatchOps.write(System.out, patch);
             
-            int version = action.dLink.append(ref, patch);
+            long version = action.dLink.append(ref, patch);
             JsonValue x = JsonNumber.value(version);
             JsonValue rslt = JsonBuilder.create()
                 .startObject()
